requirements.txt
audioread==3.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
docopt==0.6.2
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.7.0
future==1.0.0
huggingface-hub==0.33.4
idna==3.10
ImageHash==4.3.2
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
python-dotenv==1.1.1
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.0.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
urllib3==2.5.0
whisper==1.1.10
yt-dlp==2025.6.30
